Implement a Load Benchmark Runner with no flow control on scan
Due to large perfomance losses, choosing a ready channel
to send scanned batches to, when using ~100 channels
a separate implementation of the benchmark runner uses
a simplified algorithm to distirbute the scanned batches.
The algorithm follows how Prometheus distributes batches to
write to remote storage. Benchmarks show that it works better
for all kind of batch sizes. And drasticly improves performance
for large number of channels.

Also:
  - Wrote test for the scan with no flow controll
  - Added flow-control flag to tsbs_load

+++ cmd/tsbs_load_cassandra/scan.go
@@ -56,8 +56,8 @@ type eventsBatch struct {
 	rows []string
 }
 
-func (eb *eventsBatch) Len() int {
-	return len(eb.rows)
+func (eb *eventsBatch) Len() uint {
+	return uint(len(eb.rows))
 }
 
 func (eb *eventsBatch) Append(item *data.LoadedPoint) {
